Mostly fix player state
Naming conventions:
- Types(Enum/Struct) and Enum variants in UpperCamelCase
- modules in snake_case

See: https://rust-lang.github.io/api-guidelines/naming.html

Imports end in semicolons (;) and not colons(:). Arrays are defind with
[Type; length] and not [Type, length].

No clue if Wallet was supposed to be defined somewhere.

In order for a module to be compiled (and the compiler to complain), it
needs to be used or at the very least added as a module to lib.